// src/types.ts
export type Tab = {
  tabId: number;
  onlyTab: boolean;
  isOpen: boolean;
  alerts: boolean;
  count: number;
}

export type MEETING_URL_PROPERTY = {
  hash?: string;
  host?: string;
  hostname?: string;
  href?: string;
  origin?: string;
  password?: string;
  pathname?: string;
  port?: string;
  protocol?: string;
  search?: string;
  searchParams?: string;
  username?: string;
}